pyproject.toml
[project]
name = "genai-fastapi-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.9",
    "faiss-cpu==1.11.0.post1",
    "fastapi==0.116.1",
    "google-genai>=2.0.0",
    "gunicorn>=26.0.0",
    "huggingface-hub==0.34.4",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-google-genai==2.1.9",
    "openai>=2.36.0",
    "pydantic==2.11.7",
    "pymongo>=4.17.0",
    "pypdf==6.0.0",
    "python-dotenv==1.1.1",
    "python-multipart>=0.0.27",
    "sentence-transformers==5.1.0",
    "torch==2.8.0",
    "transformers==4.55.4",
    "uvicorn[standard]==0.35.0",
]
